refactor(editor): Migrate agentRequest store to stores package (no-changelog)
