+++ pnpm-workspace.yaml
@@ -52,6 +52,7 @@ catalogs:
     '@vitest/coverage-v8': ^3.1.3
     '@vitejs/plugin-vue': ^5.2.4
     '@sentry/vue': ^8.33.1
+    pinia: ^2.2.4
     typescript: ^5.8.2
     vite: ^6.3.5
     vitest: ^3.1.3
